Move ToolBox components into Panels
